Added SPEC_TESTING.md - Reduced noise when running spec tests - `--verbose` now provides better context for non-2XX responses. - Moved `linter/utils` to `_utils` because it's also being used by the tester tool (To avoid misleading stacktrace when an error appears in utils)
