+++ DEVELOPER_GUIDE.md
@@ -1,42 +1,39 @@
 <!-- TOC -->
-- [Developer Guide](#developer-guide)
-  - [Getting Started](#getting-started)
-  - [Specification](#specification)
-    - [File Structure](#file-structure)
-    - [Grouping Operations](#grouping-operations)
-    - [Grouping Schemas](#grouping-schemas)
-    - [Superseded Operations](#superseded-operations)
-    - [Global Parameters](#global-parameters)
-    - [OpenAPI Extensions](#openapi-extensions)
-  - [Writing Spec Tests](#writing-spec-tests)
-    - [Test Stories](#test-stories)
-    - [Organizing Tests](#organizing-tests)
-    - [Running Spec Tests Locally](#running-spec-tests-locally)
-  - [Tools](#tools)
-    - [Setup](#setup)
-    - [Spec Merger](#spec-merger)
-      - [Arguments](#arguments)
-      - [Example](#example)
-    - [Spec Linter](#spec-linter)
-      - [Arguments](#arguments-1)
-      - [Example](#example-1)
-    - [Dump Cluster Spec](#dump-cluster-spec)
-      - [Arguments](#arguments-2)
-      - [Example](#example-2)
-    - [Coverage](#coverage)
-      - [Arguments](#arguments-3)
-      - [Example](#example-3)
-    - [Testing](#testing)
-      - [Tests](#tests)
-      - [Lints](#lints)
-  - [Workflows](#workflows)
-    - [Analyze PR Changes](#analyze-pr-changes)
-    - [Build](#build)
-    - [Deploy GitHub Pages](#deploy-github-pages)
-    - [Comment on PR](#comment-on-pr)
-    - [Test Tools (Unit)](#test-tools-unit)
-    - [Test Tools (Integration)](#test-tools-integration)
-    - [Validate Spec](#validate-spec)
+* [Developer Guide](#developer-guide)
+  * [Getting Started](#getting-started)
+  * [Specification](#specification)
+    * [File Structure](#file-structure)
+    * [Grouping Operations](#grouping-operations)
+    * [Grouping Schemas](#grouping-schemas)
+    * [Superseded Operations](#superseded-operations)
+    * [Global Parameters](#global-parameters)
+    * [OpenAPI Extensions](#openapi-extensions)
+  * [Writing Spec Tests](#writing-spec-tests)
+  * [Tools](#tools)
+    * [Setup](#setup)
+    * [Spec Merger](#spec-merger)
+      * [Arguments](#arguments)
+      * [Example](#example)
+    * [Spec Linter](#spec-linter)
+      * [Arguments](#arguments-1)
+      * [Example](#example-1)
+    * [Dump Cluster Spec](#dump-cluster-spec)
+      * [Arguments](#arguments-2)
+      * [Example](#example-2)
+    * [Coverage](#coverage)
+      * [Arguments](#arguments-3)
+      * [Example](#example-3)
+    * [Testing](#testing)
+      * [Tests](#tests)
+      * [Lints](#lints)
+  * [Workflows](#workflows)
+    * [Analyze PR Changes](#analyze-pr-changes)
+    * [Build](#build)
+    * [Deploy GitHub Pages](#deploy-github-pages)
+    * [Comment on PR](#comment-on-pr)
+    * [Test Tools (Unit)](#test-tools--unit-)
+    * [Test Tools (Integration)](#test-tools--integration-)
+    * [Validate Spec](#validate-spec)
 <!-- TOC -->
 
 # Developer Guide
@@ -152,95 +149,7 @@ This repository includes several OpenAPI Specification Extensions to fill in any
 
 ## Writing Spec Tests
 
-To assure the correctness of the spec, you must add tests for the spec in the [tests/](tests) directory.
-
-### Test Stories
-
-Each yaml file in the tests directory represents a test story that tests a collection of related operations.
-
-A test story has 3 main components:
-- prologues: These are the operations that are executed before the test story is run. They are used to set up the environment for the test story.
-- chapters: These are the operations that are being tested.
-- epilogues: These are the operations that are executed after the test story is run. They are used to clean up the environment after the test story.
-
-Below is the simplified version of the test story that tests the [index operations](tests/indices/index.yaml):
-```yaml
-$schema: ../json_schemas/test_story.schema.yaml # The schema of the test story. Include this line so that your editor can validate the test story on the fly.
-
-description: This story tests all endpoints relevant the lifecycle of an index, from creation to deletion.
-
-prologues: [] # No prologues are needed for this story.
-
-epilogues: # Clean up the environment by assuring that the `books` index is deleted afterward.
-  - path: /books
-    method: DELETE
-    status: [200, 404] # The index may not exist, so we accept 404 as a valid response. Default to [200, 201] if not specified.
-    
-chapters:
-  - synopsis: Create an index named `books` with mappings and settings.
-    path: /{index} # The test will fail if "PUT /{index}" operation is not found in the spec.
-    method: PUT
-    parameters: # All parameters are validated against their schemas in the spec
-      index: books
-    request_body: # The request body is validated against the schema of the requestBody in the spec
-      payload:
-        mappings:
-          properties:
-            name:
-              type: keyword
-            age:
-              type: integer
-        settings:
-          number_of_shards: 5
-          number_of_replicas: 2
-    response: # The response body is validated against the schema of the corresponding response in the spec
-      status: 200 # This is the expected status code of the response. Any other status code will fail the test.
-
-  - synopsis: Retrieve the mappings and settings of the `books` index.
-    path: /{index}
-    method: GET
-    parameters:
-      index: books
-      flat_settings: true
-      
-  - synopsis: Delete the `books` index.
-    path: /{index}
-    method: DELETE
-    parameters:
-      index: books
-```
-
-Check the [test_story JSON Schema](json_schemas/test_story.schema.yaml) for the complete structure of a test story.
-
-### Organizing Tests
-
-Tests are organized in folders that match [namespaces](spec/namespaces). For example, tests for APIs defined in [spec/namespaces/indices.yaml](spec/namespaces/indices.yaml) can be found in [tests/indices/index.yaml](tests/indices/index.yaml) (for `/{index}`), and [tests/indices/_doc.yaml](tests/indices/_doc.yaml) (for `/{index}/_doc`).
-
-### Running Spec Tests Locally
-
-Set up an OpenSearch cluster with Docker using the default `OPENSEARCH_PASSWORD` (Recommended):
-```bash
-cd .github/opensearch-cluster
-docker-compose up -d
-```
-
-Run the tests (use `--opensearch-insecure` for a local cluster running in Docker that does not have a valid SSL certificate):
-```bash
-npm run test:spec -- --opensearch-insecure
-```
-
-Run a specific test story:
-```bash
-npm run test:spec -- --opensearch-insecure --tests tests/_core/info.yaml
-```
-
-If you opt to use a different password, you can set the `OPENSEARCH_PASSWORD` environment variable to the desired password before running `docker-compose up` and every time you run the tests:
-```bash
-export OPENSEARCH_PASSWORD='yourOwnPassword@2021'
-```
-
-Check the [test_story JSON Schema](json_schemas/test_story.schema.yaml) for the complete structure of a test story.
-
+To assure the correctness of the spec, you must add tests for the spec, when making changes. Check [SPECIFICATION_TESTING.md](SPECIFICATION_TESTING.md) for more information.
 
 ## Tools
